refactor(auth log): extract reusable components and fix imports
- move auth log tile and filter bar into dedicated auth_log subdirectory
- replace inline _FilterBar, _LevelChip, _LogTile with extracted components
- update all relative import paths to match new file structure
